+++ src/test/java/com/sprint/mission/discodeit/repository/ChannelRepositoryTest.java
@@ -4,10 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.fixture.ChannelFixture;
+import com.sprint.mission.discodeit.testannotation.RepositoryTest;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import com.sprint.mission.discodeit.support.RepositoryTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles("test")
